+++ src/Client.ts
@@ -8,7 +8,9 @@ import BasketLine from './Resources/Basket/BasketLine';
 import Subscriptions from './Resources/Subscriptions/Subscriptions';
 import SubscriptionAddons from './Resources/SubscriptionAddons/SubscriptionAddons';
 import Codes from './Resources/Codes/Codes';
+import Toybox from './Resources/Toybox/Toybox';
 import { loadWaccResources, WaccResources } from './Resources/Wacc';
+import Products from './Resources/Products/Products';
 
 export default class Client {
     public auth!: Auth;
@@ -19,6 +21,8 @@ export default class Client {
     public subscriptions!: Subscriptions;
     public subscriptionAddons!: SubscriptionAddons;
     public codes!: Codes;
+    public toybox!: Toybox;
+    public products!: Products;
     public wacc!: WaccResources;
 
     protected httpClient: HttpClient;
@@ -51,6 +55,8 @@ export default class Client {
         this.subscriptions = new Subscriptions(this);
         this.subscriptionAddons = new SubscriptionAddons(this);
         this.codes = new Codes(this);
+        this.toybox = new Toybox(this);
+        this.products = new Products(this);
         this.wacc = loadWaccResources(this);
     }