+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.8] - 2020-04-28
+### 🚀 Features
+- Adding resource for toybox to designate and undesignate toy for return.
+- Adding resource for products get products with query params.
+
 ## [0.0.7] - 2020-04-09
 ### 🚀 Features
 - Add method for validating reset password token.